I would use AWS for the Infrastructure platform.
The number of managed services available and integrations with IaaC tools and open-source libraries to help with scaling, monitoring and security are endless and growing by the day.

Through AWS, we can easily deploy infrastructure in a repeatable method, while designed to be scalable, fault-tolerant and secure. Our deployments can also easily integrate with our infrastructure to enable automated deployments.