The orchestration layer would be handled by Kubernetes, with a managed EKS instance in AWS.
This had scalability and portability benefits over other choices, as well as an api-first orchestration layer which is highly extensible.

This approach allowed the use of tools like Karpenter for horizontal scalability, network policies for segmentation and access control, support for mTLS between microservices, and simple PKI management.

All this is achieved while being cloud-agnostic approach of utilising Kubernetes for orchestration.